Remove credentials whenever dry-run is set from within the exchange
