+++ freqtrade/configuration/config_setup.py
@@ -3,7 +3,6 @@ from typing import Any, Dict
 
 from freqtrade.enums import RunMode
 
-from .check_exchange import remove_credentials
 from .config_validation import validate_config_consistency
 from .configuration import Configuration
 
@@ -21,8 +20,8 @@ def setup_utils_configuration(args: Dict[str, Any], method: RunMode) -> Dict[str
     configuration = Configuration(args, method)
     config = configuration.get_config()
 
-    # Ensure we do not use Exchange credentials
-    remove_credentials(config)
+    # Ensure these modes are using Dry-run
+    config['dry_run'] = True
     validate_config_consistency(config)
 
     return config
