+++ freqtrade/optimize/edge_cli.py
@@ -7,7 +7,7 @@ import logging
 from typing import Any, Dict
 
 from freqtrade import constants
-from freqtrade.configuration import TimeRange, remove_credentials, validate_config_consistency
+from freqtrade.configuration import TimeRange, validate_config_consistency
 from freqtrade.edge import Edge
 from freqtrade.optimize.optimize_reports import generate_edge_table
 from freqtrade.resolvers import ExchangeResolver, StrategyResolver
@@ -28,8 +28,8 @@ class EdgeCli:
     def __init__(self, config: Dict[str, Any]) -> None:
         self.config = config
 
-        # Reset keys for edge
-        remove_credentials(self.config)
+        # Ensure using dry-run
+        self.config['dry_run'] = True
         self.config['stake_amount'] = constants.UNLIMITED_STAKE_AMOUNT
         self.exchange = ExchangeResolver.load_exchange(self.config['exchange']['name'], self.config)
         self.strategy = StrategyResolver.load_strategy(self.config)
